iniciando functions
