+++ functions/index.js
@@ -1,20 +1,3 @@
-const functions = require("firebase-functions");
-const express = require("express");
-const cors = require("cors");
-const cookieParser = require("cookie-parser");
-const bodyParser = require("body-parser");
+const server = require("./server");
 
-const middlewares = {
-	authentication: require("./middlewares/authorization")
-};
-
-const app = express();
-app.use(cors({ origin: true }));
-app.use(bodyParser.json());
-app.use(bodyParser.urlencoded({ extended: true }));
-app.use(cookieParser());
-app.use(middlewares.authentication);
-
-app.get("/hello", (req, res) => res.json({ user: req.user }));
-
-exports.server = functions.https.onRequest(app);
+exports.server = server;
